% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\docType{data}
\name{chronicle}
\alias{chronicle}
\title{Chronicle file, 39 records}
\format{A data frame with 39 rows and 4 columns
\describe{
\item{Id_I}{Id number}
\item{Type}{Event names}
\item{Value}{Description of event}
\item{Variable}{Variable corresponding to the event}
\item{date}{Date of event}
}}
\usage{
chronicle
}
\description{
An example data set
}
\keyword{datasets}
