% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efc.R
\name{efc}
\alias{efc}
\title{Create an episodes file from a Chronicle and a VarSetup file.}
\usage{
efc(Chronicle, VarSetup)
}
\arguments{
\item{Chronicle}{The chronicle data frame.}

\item{VarSetup}{The VarSetup data frame.}
}
\value{
An episodes data frame.
}
\description{
Create an episodes file from a Chronicle and a VarSetup file.
}
