% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episodes}
\alias{episodes}
\title{Create an episodes file from a Chronicle and a Personal data frame.}
\usage{
episodes(chronicle, personal, start_event = "Birth", end_event = "Death")
}
\arguments{
\item{chronicle}{The Chronicle (time-varying) data frame.}

\item{personal}{The  personal (time-fixed) data frame.}
}
\value{
An episodes data frame.
}
\description{
Create an episodes file from a Chronicle and a Personal data frame.
}
