% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi.R
\docType{data}
\name{indi}
\alias{indi}
\title{Chronicle file with 97 persons}
\format{A data frame with 1922 rows and 7 columns
\describe{
   \item{Id_I}{Id number (character)}
   \item{Type}{Variable names}
   \item{Value}{Variable value (character)}
   \item{Year}{Year (numeric)}
   \item{Month}{Month (numeric)}
   \item{Day}{Day (numeric)}
   \item{DayFrac}{Fraction of a day (usually 0.02)}
}}
\usage{
indi
}
\description{
A training data set.
}
\keyword{datasets}
