% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varset.R
\docType{data}
\name{varset}
\alias{varset}
\title{VarSetup file}
\format{A data frame with 19 rows and 3 columns
\describe{
   \item{Type}{Id number (character)}
   \item{Transition}{Variable names}
   \item{Duration}{Variable value (character)}
}}
\usage{
varset
}
\description{
A training data set for debugging.
}
\keyword{datasets}
