% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddb.R
\docType{data}
\encoding{UTF-8}
\name{ddb}
\alias{ddb}
\title{INDIVIDUAL data  with 100 persons from DDB}
\format{A data frame with 3181 rows and 18 columns
\describe{
   \item{Id}{Record ID (int)}
   \item{Id_D}{Data base ID (int)}
   \item{Id_I}{Individual ID number (int)}
   \item{Source}{Name of register}
   \item{Type}{Variable names}
   \item{Value}{Variable value (character)}
   \item{Value_Id_C}{Id to Context}
   \item{Day}{Day (numeric)}
   \item{Month}{Month (numeric)}
   \item{Year}{Year (numeric)}
   \item{Start_day}{Start Day (numeric)}
   \item{Start_month}{Start Month (numeric)}
   \item{Start_year}{Start Year (numeric)}
   \item{End_day}{End Day (numeric)}
   \item{End_month}{End Month (numeric)}
   \item{End_year}{End Year (numeric)}
   \item{Date_type}{Date type (text)}
   \item{Missing}{Missingness (text)}
 }}
\usage{
ddb
}
\description{
A sample data set.
}
\keyword{datasets}
