% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal.R
\docType{data}
\name{personal}
\alias{personal}
\title{Personal file, 16 records}
\format{A data frame with 16 rows and 4 columns
\describe{
\item{Id_I}{Id number}
\item{Type}{Variable names}
\item{Value}{Variable value}
}}
\usage{
personal
}
\description{
An example data set
}
\keyword{datasets}
